requirements.txt
tensorflow-gpu>=1.1.0
gym
sonnet
numpy
cached_property
